3.6.1 -- Added a new tag "rg-alert"
There is still a minor issue on the default action of dismissed, a single instance of rg-alert, correctly, by default, has dissmissable set to TRUE. However, you must explicitly set it true manually for the rg-alerts list version. Syntax and features are the same as 3.5.0. Timeout works in both cases. Animations will be added later.